requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-admin-sortable2==2.2.8
django-filter==25.1
django-markdownify==0.9.5
django-ratelimit==4.1.0
djangorestframework==3.16.1
idna==3.10
Markdown==3.9
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
tinycss2==1.4.0
urllib3==2.5.0
webencodings==0.5.1
